검색창 좀 더 확대
